PCB MB-DEV update
+Added a jumper for GPIO12 for flashing and setting efuse for 3.3 Flash operation
+Added Text Variables to Schematic and PCB
+introduced versioning of pcb
+added filenames and version to schematic sheets

+++ firmware_forcelogADC/main/wifi.c
@@ -148,7 +148,7 @@ void tWifiRun(void* param)
 			}
 			else
 			{
-				ESP_LOGE(TAG_WIFI, "UNEXPECTED EVENT:0x%X", bits);
+				ESP_LOGE(TAG_WIFI, "UNEXPECTED EVENT:0x%X", (unsigned int) bits);
 			}
 
 		}
